tidy-up (#1663)
* style: make slider wider

* style: markdown simplification

* feat: add loading bar

* style: vertically center animated weather chart

Refactors the region page and weather chart components to ensure the animated chart view is properly centered on the screen.

- Adjusts the flexbox layout on the RegionPage to allow child components to fill available space.
- Modifies the WeatherChartsWithAnimation component to grow and fill the container, enabling vertical centering.

* style: temporary change

* fix: build errors

* feat: add standalone mode footer styling

Introduces a custom hook 'useIsStandalone' to detect when the application is running as an installed PWA.

This hook is used in the FooterControl component to apply specific padding, providing a better user experience in standalone mode, especially on iOS devices.

* style: tidy up card date

* fix: eslint rules

* style: make card date smaller

* style: make card date smaller

* test: update snapshot

* test: update snapshot

---------

Co-authored-by: Stuart Clark <stuart.clark@operandio.com>